qpos

+++ src/consts.rs
@@ -27,4 +27,4 @@ pub const NUM_DIGITS_GF2X_ELEMENT: usize = ((P + DIGIT_SIZE_b - 1) / DIGIT_SIZE_
 
 // Derived parameters, they are useful for QC-LDPC algorithms
 // Circulant weight structure of the Q matrix, specialized per value of N0
-pub const qBlockWeights: [[u8; 2]; 2] = [[6i32 as u8, 5i32 as u8], [5i32 as u8, 6i32 as u8]];
+pub const qBlockWeights: [[u8; 2]; 2] = [[6u8, 5u8], [5u8, 6u8]];
